requirements.txt
dataclasses==0.6
requests==2.22.0
PyYAML==5.4.1
easysnmp
influxdb
math